+++ migrations/000004_create_limit_orders_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS limit_orders (
     item_id   TEXT NOT NULL REFERENCES items(id),
     seller_id TEXT NOT NULL REFERENCES guilds(id),
     buyer_id  TEXT REFERENCES guilds(id),
-    price     NUMERIC(20,2) NOT NULL,
+    price     BIGINT NOT NULL,
     status    TEXT NOT NULL CHECK (status IN ('listed', 'sold', 'canceled')) DEFAULT 'listed',
     listed_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
